\name{setarwav-package}
\alias{setarwav-package}
\alias{setarwav}
\docType{package}
\title{
SETAR Test using Wavelet Variance and Surrogate Data
}
\description{
Code accompanying my thesis
}
\details{
\tabular{ll}{
Package: \tab setarwav\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-12-29\cr
License: \tab BSD \cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Miguel Manese <jjonphl@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
\code{\link[wmtsa:wmtsa-package]{wmtsa}}
\code{\link[tseries:tseries-package]{tseries}}
}
\examples{
~~ simple examples of the most important functions ~~
}
